Move all scene entities and systems out of core package
