+++ src/main/kotlin/no/njoh/pulseengine/modules/lighting/global/GlobalIlluminationSystem.kt
@@ -10,8 +10,8 @@ import no.njoh.pulseengine.core.graphics.api.TextureFilter.*
 import no.njoh.pulseengine.core.graphics.postprocessing.effects.MultiplyEffect
 import no.njoh.pulseengine.core.graphics.surface.SurfaceInternal
 import no.njoh.pulseengine.core.scene.SceneSystem
-import no.njoh.pulseengine.core.scene.systems.EntityRenderer
-import no.njoh.pulseengine.core.scene.systems.EntityRenderer.RenderPass
+import no.njoh.pulseengine.modules.scene.systems.EntityRenderer
+import no.njoh.pulseengine.modules.scene.systems.EntityRenderer.RenderPass
 import no.njoh.pulseengine.core.shared.annotations.Icon
 import no.njoh.pulseengine.core.shared.annotations.Name
 import no.njoh.pulseengine.core.shared.annotations.Prop
